requirements.txt
netCDF4==1.6.2
netCDF4==1.6.5
numpy==1.23.5
numpy==1.17.4
PyQt5==5.15.10
pyqtgraph==0.13.3
sip==6.8.3
wrf_python==1.3.4.1
